% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_model.R
\name{cleanup_hdr_jobs}
\alias{cleanup_hdr_jobs}
\title{Delete expired job folders}
\usage{
cleanup_hdr_jobs(root_dir, older_than_days = 14)
}
\arguments{
\item{root_dir}{Root directory containing job folders.}

\item{older_than_days}{Delete folders last modified more than this many days ago.}
}
\value{
Character vector of deleted paths, invisibly.
}
\description{
Delete expired job folders
}
