% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forgeki_aliases.R
\name{forgeKI-aliases}
\alias{forgeKI-aliases}
\alias{forgeki_config}
\alias{forgeki_guide_options}
\alias{forgeki_crisprverse_options}
\alias{forgeki_crisprverse_capabilities}
\alias{forgeki_biology_options}
\alias{forgeki_target_biology_reference_schema}
\alias{forgeki_target_biology_default_reference_path}
\alias{forgeki_load_target_biology_reference}
\alias{forgeki_build_target_biology_reference}
\alias{forgeki_build_target_biology_proteome_reference}
\alias{forgeki_fetch_uniprot_target_features}
\alias{forgeki_parse_uniprot_features}
\alias{forgeki_arm_options}
\alias{forgeki_golden_gate_options}
\alias{forgeki_mmej_options}
\alias{forgeki_stage10_options}
\alias{forgeki_runtime_options}
\alias{validate_forgeki_config}
\alias{write_forgeki_config}
\alias{run_forgeki_pipeline}
\alias{render_forgeki_report}
\alias{export_forgeki_vendor_order_sheet}
\alias{summarize_forgeki_result}
\alias{load_forgeki_result}
\alias{run_forgeki_shiny}
\alias{forgeki_default_equivalence_plan}
\alias{audit_forgeki_equivalence}
\alias{write_forgeki_equivalence_plan_template}
\alias{audit_forgeki_sequence_differences}
\alias{forgeki_normalize_order_role}
\alias{run_forgeki_stage1}
\alias{run_forgeki_stage2}
\alias{run_forgeki_stage3}
\alias{run_forgeki_stage4}
\alias{run_forgeki_stage5}
\alias{run_forgeki_stage6}
\alias{run_forgeki_stage7}
\alias{run_forgeki_stage8}
\alias{run_forgeki_stage9}
\alias{run_forgeki_stage10}
\alias{load_forgeki_cellline_reference}
\alias{load_forgeki_gene_cellline_context}
\alias{run_forgeki_stage10_gene_context}
\alias{available_forgeki_modules}
\alias{summarize_forgeki_stage8_orderability}
\title{forgeKI user-facing aliases}
\usage{
forgeki_config(...)

forgeki_guide_options(...)

forgeki_crisprverse_options(...)

forgeki_crisprverse_capabilities(...)

forgeki_biology_options(...)

forgeki_target_biology_reference_schema(...)

forgeki_target_biology_default_reference_path(...)

forgeki_load_target_biology_reference(...)

forgeki_build_target_biology_reference(...)

forgeki_build_target_biology_proteome_reference(...)

forgeki_fetch_uniprot_target_features(...)

forgeki_parse_uniprot_features(...)

forgeki_arm_options(...)

forgeki_golden_gate_options(...)

forgeki_mmej_options(...)

forgeki_stage10_options(...)

forgeki_runtime_options(...)

validate_forgeki_config(...)

write_forgeki_config(...)

run_forgeki_pipeline(...)

render_forgeki_report(...)

export_forgeki_vendor_order_sheet(...)

summarize_forgeki_result(...)

load_forgeki_result(...)

run_forgeki_shiny(...)

forgeki_default_equivalence_plan(...)

audit_forgeki_equivalence(...)

write_forgeki_equivalence_plan_template(...)

audit_forgeki_sequence_differences(...)

forgeki_normalize_order_role(...)

run_forgeki_stage1(...)

run_forgeki_stage2(...)

run_forgeki_stage3(...)

run_forgeki_stage4(...)

run_forgeki_stage5(...)

run_forgeki_stage6(...)

run_forgeki_stage7(...)

run_forgeki_stage8(...)

run_forgeki_stage9(...)

run_forgeki_stage10(...)

load_forgeki_cellline_reference(...)

load_forgeki_gene_cellline_context(...)

run_forgeki_stage10_gene_context(...)

available_forgeki_modules(...)

summarize_forgeki_stage8_orderability(...)
}
\arguments{
\item{...}{Arguments passed through to the corresponding \verb{hdr_*} function.}
}
\value{
The return value of the corresponding \verb{hdr_*} function.
}
\description{
These aliases provide forgeKI-branded entry points while preserving the stable
\verb{hdr_*} API used throughout the package internals, tests, and existing user
scripts. The \verb{hdr_*} functions remain supported; these wrappers are intended
for new user-facing examples, Shiny integration, and public-facing package
documentation.
}
