% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage1_bioc_adapter.R
\name{forgeki_hg38_bioc_packages}
\alias{forgeki_hg38_bioc_packages}
\title{List hg38 Bioconductor packages used by forgeKI}
\usage{
forgeki_hg38_bioc_packages()
}
\value{
Character vector of package names needed for default hg38 Stage 1
resource discovery.
}
\description{
List hg38 Bioconductor packages used by forgeKI
}
