% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{forgeki_select_top_designs}
\alias{forgeki_select_top_designs}
\title{Select top designs from a report model}
\usage{
forgeki_select_top_designs(model, n = 3L)
}
\arguments{
\item{model}{A \code{forgeki_report_model}.}

\item{n}{Number of rows to keep.}
}
\value{
A tibble of top designs.
}
\description{
Select top designs from a report model
}
