% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage1_bioc_adapter.R
\name{has_hdr_stage1_hg38_resources}
\alias{has_hdr_stage1_hg38_resources}
\title{Check whether hg38 Stage 1 Bioconductor resources are installed}
\usage{
has_hdr_stage1_hg38_resources()
}
\value{
\code{TRUE} if the required namespaces are available; otherwise \code{FALSE}.
}
\description{
Check whether hg38 Stage 1 Bioconductor resources are installed
}
