% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{hdr_arm_options}
\alias{hdr_arm_options}
\title{Homology-arm option defaults}
\usage{
hdr_arm_options(
  lha_target_bp = 2000L,
  rha_target_bp = 2000L,
  min_arm_bp = 300L,
  salvage_arm_bp = c(2000L, 1500L, 1000L, 650L, 500L, 450L, 350L, 300L)
)
}
\arguments{
\item{lha_target_bp}{Target upstream/left homology-arm length in base pairs.}

\item{rha_target_bp}{Target downstream/right homology-arm length in base pairs.}

\item{min_arm_bp}{Minimum accepted salvage arm length in base pairs.}

\item{salvage_arm_bp}{Ordered candidate salvage arm lengths in base pairs.}
}
\value{
A list of homology-arm options.
}
\description{
Homology-arm option defaults
}
