% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{hdr_guide_options}
\alias{hdr_guide_options}
\title{Guide option defaults}
\usage{
hdr_guide_options(
  search_radius_bp = 100L,
  top_n = 25L,
  polyt_pattern = "T{4,}",
  u6_polyt_policy = c("strict_rescue", "hard_fail", "demote", "allow"),
  polyt_demotion_multiplier = 0.1
)
}
\arguments{
\item{search_radius_bp}{Integer radius around the intended insertion site for guide discovery.}

\item{top_n}{Maximum number of guides to retain in downstream ranked outputs.}

\item{polyt_pattern}{Regular expression used to detect U6-incompatible poly-T tracts.}

\item{u6_polyt_policy}{Policy for guides with poly-T tracts.}

\item{polyt_demotion_multiplier}{Numeric multiplier applied when poly-T guides are demoted.}
}
\value{
A list of guide-design options.
}
\description{
Guide option defaults
}
