% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sequences.R
\name{hdr_reverse_complement}
\alias{hdr_reverse_complement}
\title{Reverse-complement a DNA sequence}
\usage{
hdr_reverse_complement(x)
}
\arguments{
\item{x}{DNA sequence vector.}
}
\value{
Reverse-complemented DNA sequence vector.
}
\description{
Reverse-complement a DNA sequence
}
