% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{resolve_hdr_resource}
\alias{resolve_hdr_resource}
\title{Resolve a named manifest resource}
\usage{
resolve_hdr_resource(
  manifest,
  name,
  project_dir = attr(manifest, "project_dir") \%||\% dirname(attr(manifest,
    "manifest_path"))
)
}
\arguments{
\item{manifest}{Manifest list returned by \code{read_hdr_resource_manifest()}.}

\item{name}{Resource name.}

\item{project_dir}{Project directory used to resolve relative paths.}
}
\value{
Normalized path for the requested resource.
}
\description{
Resolve a named manifest resource
}
