% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_api.R
\name{run_hdr_shiny}
\alias{run_hdr_shiny}
\title{Launch local Shiny app}
\usage{
run_hdr_shiny(...)
}
\arguments{
\item{...}{Reserved for future Shiny app options.}
}
\value{
No value in v0.0.1; raises \code{hdr_error_stage_not_implemented}.
}
\description{
Launch local Shiny app
}
