% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_hdr_config}
\alias{validate_hdr_config}
\title{Validate an HDR configuration object}
\usage{
validate_hdr_config(cfg)
}
\arguments{
\item{cfg}{Object to validate.}
}
\value{
The input \code{cfg}, invisibly, if valid.
}
\description{
Validate an HDR configuration object
}
