% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{validate_hdr_resource_manifest}
\alias{validate_hdr_resource_manifest}
\title{Validate an HDR resource manifest}
\usage{
validate_hdr_resource_manifest(
  manifest,
  project_dir = attr(manifest, "project_dir") \%||\% dirname(attr(manifest,
    "manifest_path")),
  required = character()
)
}
\arguments{
\item{manifest}{Manifest list.}

\item{project_dir}{Project directory used to resolve relative paths.}

\item{required}{Character vector of required resource names.}
}
\value{
An \code{hdr_resource_status} tibble.
}
\description{
Validate an HDR resource manifest
}
