% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmej_cellline_reference.R
\name{validate_mmej_cellline_reference}
\alias{validate_mmej_cellline_reference}
\alias{hdr_validate_mmej_cellline_reference}
\title{Validate a standardized MMEJ cell-line reference}
\usage{
validate_mmej_cellline_reference(x)

hdr_validate_mmej_cellline_reference(x)
}
\arguments{
\item{x}{Data frame or standardized reference object.}
}
\value{
A tibble with validation checks.
}
\description{
Validate a standardized MMEJ cell-line reference
}
