% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_model.R
\name{zip_hdr_job_outputs}
\alias{zip_hdr_job_outputs}
\title{Zip HDR job outputs}
\usage{
zip_hdr_job_outputs(
  job,
  zip_path = file.path(job$job_dir, "outputs", "full_output_bundle.zip")
)
}
\arguments{
\item{job}{HDR job object returned by \code{new_hdr_job()}.}

\item{zip_path}{Output zip-file path.}
}
\value{
Normalized zip-file path.
}
\description{
Zip HDR job outputs
}
